Extract input, output and options
