+++ src/main/scala/zio/redis/Output.scala
@@ -1,0 +1,18 @@
+package zio.redis
+
+import zio.{ Chunk, IO }
+import zio.stream.Stream
+import zio.duration.Duration
+
+sealed trait Output[+A]
+
+object Output {
+  case object BoolOutput     extends Output[IO[Error, Boolean]]
+  case object ByteOutput     extends Output[IO[Error, Chunk[Byte]]]
+  case object DurationOutput extends Output[IO[Error, Duration]]
+  case object LongOutput     extends Output[IO[Error, Long]]
+  case object ScanOutput     extends Output[IO[Error, (Long, Stream[Error, Chunk[Byte]])]]
+  case object StreamOutput   extends Output[Stream[Error, Chunk[Byte]]]
+  case object StringOutput   extends Output[IO[Error, String]]
+  case object UnitOutput     extends Output[IO[Error, Unit]]
+}
