+++ src/main/scala/zio/redis/api/Connection.scala
@@ -1,8 +1,8 @@
 package zio.redis.api
 
 import zio.redis.Command
-import zio.redis.Command.Input._
-import zio.redis.Command.Output._
+import zio.redis.Input._
+import zio.redis.Output._
 
 trait Connection {
   final val auth   = Command("AUTH", StringInput, UnitOutput)
