+++ src/main/scala/zio/redis/api/SortedSets.scala
@@ -1,13 +1,16 @@
 package zio.redis.api
 
 import zio.redis.Command
-import zio.redis.Command.Input._
-import zio.redis.Command.Output._
+import zio.redis.Input._
+import zio.redis.Output._
 
 trait SortedSets {
+  final val bzpopmin = Command("BZPOPMIN", Tuple2(DurationInput, NonEmptyList(StringInput)), ByteOutput)
+  final val bzpopmax = Command("BZPOPMAX", Tuple2(DurationInput, NonEmptyList(StringInput)), ByteOutput)
   final val zcard    = Command("ZCARD", StringInput, LongOutput)
   final val zcount   = Command("ZCOUNT", Tuple2(StringInput, RangeInput), LongOutput)
   final val zincrby  = Command("ZINCRBY", Tuple3(StringInput, LongInput, ByteInput), LongOutput)
+  final val zrange   = Command("ZRANGE", Tuple3(StringInput, RangeInput, OptionalInput(WithScoresInput)), StreamOutput)
   final val zrank    = Command("ZRANK", Tuple2(StringInput, ByteInput), ByteOutput)
   final val zrem     = Command("ZREM", Tuple2(StringInput, NonEmptyList(ByteInput)), LongOutput)
   final val zrevrank = Command("ZREVRANK", Tuple2(StringInput, ByteInput), LongOutput)
