+++ src/main/scala/zio/redis/options/WithScores.scala
@@ -1,0 +1,3 @@
+package zio.redis.options
+
+case object WithScores
